requirements.txt
# Streamlit app
streamlit>=1.33.0
python-dotenv>=1.0.1

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7

# YouTube transcript extraction
youtube-transcript-api>=0.6.2

# Vector database (FAISS)
faiss-cpu>=1.7.4

# OpenAI SDK
openai>=1.30.1

# Optional: helpful for dev/debug
tqdm>=4.66.4
